safevault_pro/requirements.txt
customtkinter>=5.2.0
cryptography>=41.0.0
pyperclip>=1.8.0
pillow>=10.0.0
